pip_requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0
django-extensions==2.2.5
django-filter==2.2.0
django-guardian==2.2.0
djangorestframework==3.11.0
idna==2.8
Pillow==6.2.1
pkg-resources==0.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.8
uuid==1.30
